+++ backend/db/models.py
@@ -6,7 +6,7 @@ ERD 기반으로 SQLAlchemy 모델 생성
 from sqlalchemy import Column, Integer, String, DateTime, Date, Enum, ForeignKey, Float, Text, TIMESTAMP
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
-from database import Base
+from db.database import Base
 import enum
 
 # 성별 Enum 정의
